+++ packages/client/src/clients/guide/index.ts
@@ -4,6 +4,7 @@ export {
   checkActivatable,
 } from "./client";
 export { checkStateIfThrottled } from "./helpers";
+export type { ToolbarV2RunConfig } from "./helpers";
 export type {
   KnockGuide,
   KnockGuideStep,