+++ examples/expo-example/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @knocklabs/expo-example
 
+## 1.1.1
+
+### Patch Changes
+
+- @knocklabs/expo@0.6.1
+
 ## 1.1.0
 
 ### Minor Changes